requirements.txt
pyserial==3.4
requests=2.21.0
pytz=2019.3
